Fixing bug where tiled editor supports decimals for object height but ntiled fails to parse it.

+++ src/NTiled/Domain/TiledObject.cs
@@ -65,7 +65,7 @@ namespace NTiled
         /// Gets or sets the height of the object in pixels.
         /// </summary>
         /// <value>The height of the object in pixels.</value>
-        public int Height { get; set; }
+        public decimal Height { get; set; }
 
         /// <summary>
         /// Gets the object properties.
